refactor connections
