+++ app/src/main/AndroidManifest.xml
@@ -229,7 +229,7 @@
             </intent-filter>
         </activity>
         <service
-            android:name=".data.connection.discord.DiscordRPCService"
+            android:name=".data.connections.discord.DiscordRPCService"
             android:enabled="true"
             android:exported="false"
             android:foregroundServiceType="dataSync">
