+++ app/src/main/java/eu/kanade/presentation/more/settings/Preference.kt
@@ -4,7 +4,8 @@ import androidx.annotation.IntRange
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.graphics.vector.ImageVector
-import eu.kanade.tachiyomi.data.connection.BaseConnection
+import eu.kanade.core.preference.asState
+import eu.kanade.tachiyomi.data.connections.ConnectionsService
 import eu.kanade.tachiyomi.data.track.Tracker
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.ImmutableMap
@@ -160,7 +161,7 @@ sealed class Preference {
          * A [PreferenceItem] for individual connections service.
          */
         data class ConnectionPreference(
-            val service: BaseConnection,
+            val service: ConnectionsService,
             override val title: String,
             val login: () -> Unit,
             val openSettings: () -> Unit,
