+++ app/src/main/java/eu/kanade/tachiyomi/di/AppModule.kt
@@ -16,7 +16,7 @@ import eu.kanade.tachiyomi.data.cache.ChapterCache
 import eu.kanade.tachiyomi.data.cache.CoverCache
 import eu.kanade.tachiyomi.data.cache.PagePreviewCache
 import eu.kanade.tachiyomi.BuildConfig
-import eu.kanade.tachiyomi.data.connection.ConnectionManager
+import eu.kanade.tachiyomi.data.connections.ConnectionsManager
 import eu.kanade.tachiyomi.data.download.DownloadCache
 import eu.kanade.tachiyomi.data.download.DownloadManager
 import eu.kanade.tachiyomi.data.download.DownloadProvider
@@ -184,7 +184,7 @@ class AppModule(val app: Application) : InjektModule {
         // KMK <--
 
         // AM (CONNECTIONS) -->
-        addSingletonFactory { ConnectionManager() }
+        addSingletonFactory { ConnectionsManager() }
         // <-- AM (CONNECTIONS)
 
         // Asynchronously init expensive components for a faster cold start
